requirements.txt
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.2
aiohttp==3.9.1
asyncio==3.4.3
pytest==7.4.3
mypy==1.7.1 